generator/requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0

# Document generation
jinja2==3.1.2
markdown==3.5.1
weasyprint==60.1  # For PDF generation
python-docx==1.1.0  # For Word documents

# Message Queue
aio-pika==9.3.0

# Storage
minio==7.2.0

# Utils
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1